refactor(database): move init out of the data-view
